requirements.txt
PyYAML==6.0.1
ewatercycle~=2.1.0
ewatercycle-hype==0.0.1
ewatercycle-leakybucket==0.0.2
ewatercycle-lisflood==0.0.1
ewatercycle-marrmot==0.0.2
ewatercycle-pcrglobwb==0.0.2
ewatercycle-wflow==0.0.2
matplotlib~=3.8.4
markdown2~=2.4.13
Cython~=3.0.10
netCDF4~=1.6.5
numpy~=1.26.4
pytest~=8.2.0
ttkbootstrap~=1.10.1
xarray~=2024.5.0
era5cli==1.4.2
#pylint==3.2.3
#isort==5.13.2
